Add files via upload
How to create an API key. I know it kind of sucks, but you do need to upload some kind of credit card information into Google. I can guarantee you that in the 3 years I have been playing with this I have not paid a single cent! There is absolutely nothing to be afraid of unless you use more than 1GB of map data a day, which is equivalent to more than 20.000 map requests A. DAY. I don't care how much you love ggmap, because it sure as hell isn't that much.